+++ src/platforms/swlink/usbdfu.c
@@ -26,7 +26,7 @@
 #include "usbdfu.h"
 #include "platform.h"
 
-uint32_t app_address = 0x08002000;
+uintptr_t app_address = 0x08002000;
 uint32_t rev;
 
 void dfu_detach(void)
